+++ src/main/java/mod_RorysExcavation.java
@@ -19,7 +19,7 @@ import java.io.File;
 public class mod_RorysExcavation extends BaseMod {
 
     public static final String MOD_NAME = "Rory's Excavation";
-    public static final String VERSION  = "0.4.0";
+    public static final String VERSION  = "0.5.0";
 
     @Override
     public String getVersion() {